Added split on key + improved dictionary
Improved the way a dictionary is handled a lot, so it is now much
simpler to add more dictionary watchers.

TODO: add something along the lines of a watcher class, that does
has a function to get a value every tick, keeps the current value
and old value, and splits or does whatever if needed.